Move to hooks.
